Work on dark pool short menu
