+++ gamestonk_terminal/stocks/dark_pool_shorts/__init__.py
@@ -1,0 +1,1 @@
+from . import dps_controller  # noqa: F401
