+++ gamestonk_terminal/stocks/dark_pool_shorts/finra_model.py
@@ -1,14 +1,10 @@
-""" FINRA View """
+""" FINRA Model """
 __docformat__ = "numpy"
 
-import argparse
-from typing import List, Tuple
+from typing import List, Tuple, Dict
 import requests
 import pandas as pd
-from matplotlib import pyplot as plt
-from gamestonk_terminal.config_plot import PLOT_DPI
-from gamestonk_terminal import feature_flags as gtff
-from gamestonk_terminal.helper_funcs import parse_known_args_and_warn, plot_autoscale
+from scipy import stats
 
 
 def getFINRAweeks(tier, is_ats) -> List:
@@ -55,6 +51,55 @@ def getFINRAweeks(tier, is_ats) -> List:
     )
 
     return response.json() if response.status_code == 200 else list()
+
+
+def getFINRAdata_offset(weekStartDate, tier, ticker, is_ats, offset):
+    req_hdr = {"Accept": "application/json", "Content-Type": "application/json"}
+
+    l_cmp_filters = [
+        {
+            "compareType": "EQUAL",
+            "fieldName": "weekStartDate",
+            "fieldValue": weekStartDate,
+        },
+        {"compareType": "EQUAL", "fieldName": "tierIdentifier", "fieldValue": tier},
+        {
+            "compareType": "EQUAL",
+            "description": "",
+            "fieldName": "summaryTypeCode",
+            "fieldValue": "ATS_W_SMBL" if is_ats else "OTC_W_SMBL",
+        },
+    ]
+
+    if ticker:
+        l_cmp_filters.append(
+            {
+                "compareType": "EQUAL",
+                "fieldName": "issueSymbolIdentifier",
+                "fieldValue": ticker,
+            }
+        )
+
+    req_data = {
+        "compareFilters": l_cmp_filters,
+        "delimiter": "|",
+        "fields": [
+            "issueSymbolIdentifier",
+            "totalWeeklyShareQuantity",
+            "totalWeeklyTradeCount",
+            "lastUpdateDate",
+        ],
+        "limit": 5000,
+        "offset": offset,
+        "quoteValues": False,
+        "sortFields": ["totalWeeklyShareQuantity"],
+    }
+
+    return requests.post(
+        "https://api.finra.org/data/group/otcMarket/name/weeklySummary",
+        headers=req_hdr,
+        json=req_data,
+    )
 
 
 def getFINRAdata(
@@ -132,6 +177,76 @@ def getFINRAdata(
     )
 
 
+def getATSdata(num_tickers_to_filter: int) -> Tuple[pd.DataFrame, Dict]:
+    """Get all FINRA ATS data, and parse most promising tickers based on linear regression
+
+    Parameters
+    ----------
+    num_tickers_to_filter : int
+        Number of tickers to filter from entire ATS data based on the sum of the total weekly shares quantity
+
+    Returns
+    ----------
+    pd.DataFrame
+        Dark Pools (ATS) Data
+    Dict
+        Tickers from Dark Pools with better regression slope
+    """
+    tiers = ["T1", "T2", "OTCE"]
+    df_ats = pd.DataFrame()
+
+    for tier in tiers:
+        print(f"Processing Tier {tier} ...")
+        for d_week in getFINRAweeks(tier, is_ats=True):
+            offset = 0
+            response = getFINRAdata_offset(
+                d_week["weekStartDate"], tier, "", True, offset
+            )
+            l_data = response.json()
+
+            while len(response.json()) == 5000:
+                offset += 5000
+                response = getFINRAdata_offset(
+                    d_week["weekStartDate"], tier, "", True, offset
+                )
+                l_data += response.json()
+
+            df_ats_week = pd.DataFrame(l_data)
+            df_ats_week["weekStartDate"] = d_week["weekStartDate"]
+
+            if not df_ats_week.empty:
+                df_ats = df_ats.append(df_ats_week, ignore_index=True)
+
+    df_ats = df_ats.sort_values("weekStartDate")
+    df_ats["weekStartDateInt"] = pd.to_datetime(df_ats["weekStartDate"]).apply(
+        lambda x: x.timestamp()
+    )
+
+    print(f"Processing regression on {num_tickers_to_filter} promising tickers ...")
+
+    d_ats_reg = {}
+    # set(df_ats['issueSymbolIdentifier'].values) this would be iterating through all tickers
+    # but that is extremely time consuming for little reward. A little filtering is done to
+    # speed up search for best ATS tickers
+    for symbol in list(
+        df_ats.groupby("issueSymbolIdentifier")["totalWeeklyShareQuantity"]
+        .sum()
+        .sort_values()[-num_tickers_to_filter:]
+        .index
+    ):
+        slope = stats.linregress(
+            df_ats[df_ats["issueSymbolIdentifier"] == symbol][
+                "weekStartDateInt"
+            ].values,
+            df_ats[df_ats["issueSymbolIdentifier"] == symbol][
+                "totalWeeklyShareQuantity"
+            ].values,
+        )[0]
+        d_ats_reg[symbol] = slope
+
+    return df_ats, d_ats_reg
+
+
 def getTickerFINRAdata(ticker: str) -> Tuple[pd.DataFrame, pd.DataFrame]:
     """Get all FINRA data associated with a ticker
 
@@ -188,117 +303,3 @@ def getTickerFINRAdata(ticker: str) -> Tuple[pd.DataFrame, pd.DataFrame]:
         df_otc = df_otc.set_index("weekStartDate")
 
     return df_ats, df_otc
-
-
-def plot_dark_pools(ticker: str, ats: pd.DataFrame, otc: pd.DataFrame):
-    """Plots ATS and NON-ATS data
-
-    Parameters
-    ----------
-    ticker : str
-        Stock ticker to get data from
-    ats : pd.DataFrame
-        Dark Pools (ATS) Data
-    otc : pd.DataFrame
-        OTC (Non-ATS) Data
-    """
-    _, _ = plt.subplots(figsize=plot_autoscale(), dpi=PLOT_DPI)
-
-    plt.subplot(3, 1, (1, 2))
-    if not ats.empty and not otc.empty:
-        plt.bar(
-            ats.index,
-            (ats["totalWeeklyShareQuantity"] + otc["totalWeeklyShareQuantity"])
-            / 1_000_000,
-            color="tab:orange",
-        )
-        plt.bar(
-            otc.index, otc["totalWeeklyShareQuantity"] / 1_000_000, color="tab:blue"
-        )
-        plt.legend(["ATS", "OTC"])
-
-    elif not ats.empty:
-        plt.bar(
-            ats.index,
-            ats["totalWeeklyShareQuantity"] / 1_000_000,
-            color="tab:orange",
-        )
-        plt.legend(["ATS"])
-
-    elif not otc.empty:
-        plt.bar(
-            otc.index, otc["totalWeeklyShareQuantity"] / 1_000_000, color="tab:blue"
-        )
-        plt.legend(["OTC"])
-
-    plt.ylabel("Total Weekly Shares [Million]")
-    plt.grid(b=True, which="major", color="#666666", linestyle="-", alpha=0.2)
-    plt.title(f"Dark Pools (ATS) vs OTC (Non-ATS) Data for {ticker}")
-
-    plt.subplot(313)
-    if not ats.empty:
-        plt.plot(
-            ats.index,
-            ats["totalWeeklyShareQuantity"] / ats["totalWeeklyTradeCount"],
-            color="tab:orange",
-        )
-        plt.legend(["ATS"])
-
-        if not otc.empty:
-            plt.plot(
-                otc.index,
-                otc["totalWeeklyShareQuantity"] / otc["totalWeeklyTradeCount"],
-                color="tab:blue",
-            )
-            plt.legend(["ATS", "OTC"])
-
-    else:
-        plt.plot(
-            otc.index,
-            otc["totalWeeklyShareQuantity"] / otc["totalWeeklyTradeCount"],
-            color="tab:blue",
-        )
-        plt.legend(["OTC"])
-
-    plt.ylabel("Shares per Trade")
-    plt.grid(b=True, which="major", color="#666666", linestyle="-", alpha=0.2)
-    plt.gcf().autofmt_xdate()
-    plt.xlabel("Weeks")
-
-    if gtff.USE_ION:
-        plt.ion()
-
-    plt.show()
-
-
-def dark_pool(other_args: List[str], ticker: str):
-    """Display barchart of dark pool (ATS) and OTC (Non ATS) data
-
-    Parameters
-    ----------
-    other_args : List[str]
-        argparse other args
-    ticker : str
-        Stock ticker
-    """
-    parser = argparse.ArgumentParser(
-        add_help=False,
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        prog="dp",
-        description="Display barchart of dark pool (ATS) and OTC (Non ATS) data [Source: FINRA]",
-    )
-    try:
-        ns_parser = parse_known_args_and_warn(parser, other_args)
-        if not ns_parser:
-            return
-
-        df_ats, df_otc = getTickerFINRAdata(ticker)
-
-        if df_ats.empty and df_otc.empty:
-            print("No ticker data found!")
-
-        plot_dark_pools(ticker, df_ats, df_otc)
-        print("")
-
-    except Exception as e:
-        print(e, "\n")
